build: Rewrites project build definition. #SW-159
This patch rewrites the definition of the build of libqi-core in order to
remove the dependency to `qibuild` and `qi toolchains` and instead use
more standard tools and conventions.

It completely refactors the `CMakeLists.txt` files to use an approach
closer to what is referred to as "modern CMake". The project now requires
CMake 3.23 or greater.